consolidate abstractive summarizers
